requirements.txt
torch
torchvision
facenet-pytorch
lpips
numpy
pandas
pillow
opencv-python
openpyxl
fer
deepface
tf-keras
